requirements.txt
python==3.11.*
datacompy==0.11.0
pandas==2.1.4
pyspark==3.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
cx-Oracle==8.3.0
psycopg2-binary==2.9.9
pyodbc==5.0.1
requests==2.31.0
concurrent-log-handler==0.9.24
APScheduler==3.10.4
click==8.1.7
jinja2==3.1.2
pycryptodome==3.19.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0